Switch from telescope to fzf.
I ran into a few issues with telescope:

- It was slow.
- It had a number of bugs, and their bug tracker requires a herculean
  effort to submit a new bug.

So this drops it in favor of fzf, which I've used previously with
success.

I also added ripgrep, or `rg`. In the past I've used `ag`, but I've
heard good things about `rg` so it felt like a good time to give
it a shot.

I also fixed a bug in `config.fish`. The `PATH` munging I was doing
was actually changing the working directory when launching a new
shell, which was causing a lot of weird side-effects. Oops.